REFACT: Move code from main window to settings and config modules
