+++ src/darsia/gui/ui/settings.py
@@ -1,5 +1,7 @@
 """Settings and input widget factory for DarSIA GUI."""
 
+import ast
+
 from PySide6.QtWidgets import (
     QCheckBox,
     QComboBox,
@@ -2232,3 +2234,488 @@ class SettingsFactory:
         result["widget"] = group_box
         result["is_group_result"] = True
         return None, result
+
+    def _sync_settings_inputs_to_config_dict(self):
+        """Flush all live settings widget values into main_window.config_dict.
+
+        Shared by save_settings() (before writing to disk) and
+        _render_settings_tabs() (before it deletes the current widgets and
+        resets settings_inputs), so in-progress edits are never silently
+        discarded by navigating between tabs before saving.
+        """
+        import re
+        from PySide6.QtWidgets import QLineEdit, QComboBox, QCheckBox
+
+        # First pass: collect group checkbox states and determine which sub-inputs to skip
+        group_active_names: dict[str, set[str]] = {}
+        skip_keys: set[str] = set()
+
+        for key, value in self.main_window.settings_inputs.items():
+            if isinstance(value, dict) and "checkbox" in value:
+                # Handle section-level boolean toggle (active_bool_key)
+                if "bool_key" in value:
+                    self.set_value(
+                        self.main_window.config_dict,
+                        value["bool_key"],
+                        value["checkbox"].isChecked(),
+                    )
+                    if not value["checkbox"].isChecked():
+                        skip_keys.update(value["sub_inputs"].keys())
+                    continue
+
+                # Handle field-level toggles within a section (active_list_key)
+                # This is a group dict; record its active state
+                active_list_key = value["active_list_key"]
+                group_active_names.setdefault(active_list_key, set())
+                if value["checkbox"].isChecked():
+                    group_active_names[active_list_key].add(value["name"])
+                else:
+                    # Group is unchecked; skip saving any of its sub-inputs
+                    skip_keys.update(value["sub_inputs"].keys())
+
+        # Second pass: save all regular values (non-group dicts),
+        # skipping unchecked group sub-inputs.
+        for key, value in self.main_window.settings_inputs.items():
+            # Skip group result dicts (marked with is_group_result)
+            if isinstance(value, dict) and value.get("is_group_result"):
+                continue
+            # Skip group dicts with checkboxes (already handled above)
+            if isinstance(value, dict) and "checkbox" in value:
+                continue
+            # Skip path_map, int_group_list, int_list_map, time_interval_map, time_window_map,
+            # time_data_map, path_data_map, registry_key_list, format_map dicts (handled below)
+            if isinstance(value, dict) and (
+                "path_map" in value
+                or "int_group_list" in value
+                or "int_list_map" in value
+                or "time_interval_map" in value
+                or "time_window_map" in value
+                or "time_data_map" in value
+                or "path_data_map" in value
+                or "registry_key_list" in value
+                or "format_map" in value
+                or "format_key_list" in value
+            ):
+                continue
+            # Skip sub-inputs of unchecked groups
+            if key in skip_keys:
+                continue
+
+            # Unwrap composite field widgets (wrapper widget containing type label +
+            # help button) to extract the actual editable control (QLineEdit, QComboBox,
+            # QCheckBox)
+            value = unwrap_composite_widget(value)
+
+            try:
+                if isinstance(value, QLineEdit):
+                    self.set_value(
+                        self.main_window.config_dict, key, ast.literal_eval(value.text())
+                    )
+                elif isinstance(value, QComboBox):
+                    self.set_value(
+                        self.main_window.config_dict, key, value.currentText()
+                    )
+                elif isinstance(value, QCheckBox):
+                    self.set_value(
+                        self.main_window.config_dict, key, value.isChecked()
+                    )
+                elif isinstance(value, list):
+                    if len(value) > 0:
+                        if isinstance(value[0], QCheckBox):
+                            self.set_value(
+                                self.main_window.config_dict,
+                                key,
+                                [item.text() for item in value if item.isChecked()],
+                            )
+                        elif isinstance(value[0], QLineEdit):
+                            self.set_value(
+                                self.main_window.config_dict,
+                                key,
+                                [item.text() for item in value if item.text().strip()],
+                            )
+            except (ValueError, SyntaxError):
+                if hasattr(value, "text"):
+                    self.set_value(self.main_window.config_dict, key, value.text())
+
+        # Third pass: save path_map dicts (key -> value mappings)
+        for key, value in self.main_window.settings_inputs.items():
+            if isinstance(value, dict) and "path_map" in value:
+                rows = value["rows"]
+                result = {
+                    k.text(): v.text()
+                    for k, v in rows
+                    if k.text().strip() and v.text().strip()
+                }
+                self.set_value(self.main_window.config_dict, key, result)
+
+        # Fourth pass: parse int_group_list rows into list[list[int]]
+        for key, value in self.main_window.settings_inputs.items():
+            if isinstance(value, dict) and "int_group_list" in value:
+                groups = []
+                for edit in value["rows"]:
+                    text = edit.text().strip()
+                    if not text:
+                        continue
+                    tokens = [t for t in re.split(r"[,\s]+", text) if t]
+                    try:
+                        groups.append([int(t) for t in tokens])
+                    except ValueError:
+                        self.main_window.print_log(
+                            f"Skipping invalid group '{text}' for {key}: not all-integer."
+                        )
+                self.set_value(self.main_window.config_dict, key, groups)
+
+        # Fifth pass: parse int_list_map rows into dict[int, list[int]]
+        for key, value in self.main_window.settings_inputs.items():
+            if isinstance(value, dict) and "int_list_map" in value:
+                result = {}
+                for key_edit, value_edit in value["rows"]:
+                    key_text = key_edit.text().strip()
+                    value_text = value_edit.text().strip()
+                    if not key_text or not value_text:
+                        continue
+                    tokens = [t for t in re.split(r"[,\s]+", value_text) if t]
+                    try:
+                        result[int(key_text)] = [int(t) for t in tokens]
+                    except ValueError:
+                        self.main_window.print_log(
+                            f"Skipping invalid row '{key_text}: {value_text}' for {key}: "
+                            "key and values must all be integers."
+                        )
+                if value.get("flatten_in_section"):
+                    section = value["section"]
+                    section_dict = self.main_window.config_dict.setdefault(section, {})
+                    # Drop id sub-tables removed in the GUI (has "labels", not in new result)
+                    stale_ids = [
+                        k
+                        for k, v in section_dict.items()
+                        if isinstance(v, dict)
+                        and "labels" in v
+                        and k not in {str(i) for i in result}
+                    ]
+                    for k in stale_ids:
+                        del section_dict[k]
+                    for facies_id, labels in result.items():
+                        section_dict[str(facies_id)] = {"labels": labels}
+                else:
+                    self.set_value(self.main_window.config_dict, key, result)
+
+        # Sixth pass: parse time_interval_map rows into dict[str, {start/end/num/tol}]
+        for key, value in self.main_window.settings_inputs.items():
+            if isinstance(value, dict) and "time_interval_map" in value:
+                result = {}
+                for name_edit, start_edit, end_edit, num_edit, tol_edit in value[
+                    "rows"
+                ]:
+                    name_text = name_edit.text().strip()
+                    if not name_text:
+                        continue
+                    result[name_text] = {
+                        "start": start_edit.text().strip() or "00:00:00",
+                        "end": end_edit.text().strip() or "00:00:00",
+                        "num": (
+                            int(num_edit.text().strip())
+                            if num_edit.text().strip().isdigit()
+                            else 0
+                        ),
+                        "tol": tol_edit.text().strip() or "00:00:00",
+                    }
+                self.set_value(self.main_window.config_dict, key, result)
+
+        # Seventh pass: parse time_window_map rows into dict[str, {start/end}]
+        for key, value in self.main_window.settings_inputs.items():
+            if isinstance(value, dict) and "time_window_map" in value:
+                result = {}
+                for name_edit, start_edit, end_edit in value["rows"]:
+                    name_text = name_edit.text().strip()
+                    if not name_text:
+                        continue
+                    result[name_text] = {
+                        "start": start_edit.text().strip() or "00:00:00",
+                        "end": end_edit.text().strip() or "00:00:00",
+                    }
+                self.set_value(self.main_window.config_dict, key, result)
+
+        # Eighth pass: parse time_data_map rows into dict[str, {times}]
+        for key, value in self.main_window.settings_inputs.items():
+            if isinstance(value, dict) and "time_data_map" in value:
+                result = {}
+                for name_edit, times_edit in value["rows"]:
+                    name_text = name_edit.text().strip()
+                    if not name_text:
+                        continue
+                    times_text = times_edit.text().strip()
+                    if times_text:
+                        # Parse comma-separated times
+                        times_list = [
+                            t.strip() for t in times_text.split(",") if t.strip()
+                        ]
+                        result[name_text] = {"times": times_list}
+                    else:
+                        result[name_text] = {"times": []}
+                self.set_value(self.main_window.config_dict, key, result)
+
+        # Ninth pass: parse path_data_map rows into dict[str, {paths}]
+        for key, value in self.main_window.settings_inputs.items():
+            if isinstance(value, dict) and "path_data_map" in value:
+                result = {}
+                for name_edit, paths_edit in value["rows"]:
+                    name_text = name_edit.text().strip()
+                    if not name_text:
+                        continue
+                    paths_text = paths_edit.text().strip()
+                    if paths_text:
+                        # Parse comma-separated paths
+                        paths_list = [
+                            p.strip() for p in paths_text.split(",") if p.strip()
+                        ]
+                        result[name_text] = {"paths": paths_list}
+                    else:
+                        result[name_text] = {"paths": []}
+                self.set_value(self.main_window.config_dict, key, result)
+
+        # Tenth pass: parse registry_key_list rows into list[str]
+        for key, value in self.main_window.settings_inputs.items():
+            if isinstance(value, dict) and "registry_key_list" in value:
+                result = []
+                seen = set()
+                for combo in value["rows"]:
+                    text = combo.currentText().strip()
+                    if text and text not in seen:
+                        result.append(text)
+                        seen.add(text)
+                # Write as list[str] (empty list if no selections, never None)
+                if result:
+                    self.set_value(self.main_window.config_dict, key, result)
+                else:
+                    # Empty selection: delete the key (or set to empty list)
+                    # For now, set to empty list to match the field's Optional nature
+                    self.set_value(self.main_window.config_dict, key, None)
+
+        # Eleventh pass: parse format_map rows into list[dict] for [[format]] TOML shape
+        for key, value in self.main_window.settings_inputs.items():
+            if isinstance(value, dict) and "format_map" in value:
+                result = []
+                for (
+                    name_edit,
+                    type_combo,
+                    filename_pattern_combo,
+                    resolution_edit,
+                    keep_ratio_check,
+                    dpi_edit,
+                    cmap_edit,
+                    quality_edit,
+                    compression_edit,
+                    dtype_edit,
+                    delimiter_edit,
+                    header_edit,
+                    float_format_edit,
+                ) in value["rows"]:
+                    name_text = name_edit.text().strip()
+                    if not name_text:
+                        continue
+
+                    # Always-required fields
+                    entry = {
+                        "type": type_combo.currentText().strip(),
+                        "name": name_text,
+                        "filename_pattern": filename_pattern_combo.currentText().strip(),
+                    }
+
+                    # Optional fields: empty text means None/omit
+                    # resolution: comma-separated rows,cols
+                    res_text = resolution_edit.text().strip()
+                    if res_text:
+                        parts = [p.strip() for p in res_text.split(",")]
+                        if len(parts) == 2:
+                            try:
+                                entry["resolution"] = [int(parts[0]), int(parts[1])]
+                            except ValueError:
+                                pass
+
+                    # keep_ratio (bool)
+                    if keep_ratio_check.isChecked():
+                        entry["keep_ratio"] = True
+
+                    # Type-specific optional fields
+                    dpi_text = dpi_edit.text().strip()
+                    if dpi_text:
+                        try:
+                            entry["dpi"] = int(dpi_text)
+                        except ValueError:
+                            pass
+
+                    cmap_text = cmap_edit.text().strip()
+                    if cmap_text:
+                        entry["cmap"] = cmap_text
+
+                    quality_text = quality_edit.text().strip()
+                    if quality_text:
+                        try:
+                            entry["quality"] = int(quality_text)
+                        except ValueError:
+                            pass
+
+                    compression_text = compression_edit.text().strip()
+                    if compression_text:
+                        try:
+                            entry["compression"] = int(compression_text)
+                        except ValueError:
+                            pass
+
+                    dtype_text = dtype_edit.text().strip()
+                    if dtype_text:
+                        entry["dtype"] = dtype_text
+
+                    # CSV-specific fields
+                    delimiter_text = delimiter_edit.text().strip()
+                    if delimiter_text and delimiter_text != ",":
+                        entry["delimiter"] = delimiter_text
+
+                    header_text = header_edit.text().strip()
+                    if header_text:
+                        entry["header"] = header_text
+
+                    float_format_text = float_format_edit.text().strip()
+                    if float_format_text and float_format_text != "{:.2e}":
+                        entry["float_format"] = float_format_text
+
+                    result.append(entry)
+
+                # Write as list[dict] directly to config_dict["format"]
+                self.main_window.config_dict["format"] = result
+
+        # Twelfth pass: parse format_key_list rows into list[str] (or single str)
+        for key, value in self.main_window.settings_inputs.items():
+            if isinstance(value, dict) and "format_key_list" in value:
+                result = []
+                seen = set()
+                for combo in value["rows"]:
+                    text = combo.currentText().strip()
+                    if text and text not in seen:
+                        result.append(text)
+                        seen.add(text)
+                if value.get("max_rows") == 1:
+                    self.set_value(
+                        self.main_window.config_dict, key, result[0] if result else None
+                    )
+                else:
+                    self.set_value(
+                        self.main_window.config_dict, key, result if result else None
+                    )
+
+        # Thirteenth pass: write all active lists
+        for active_list_key, names in group_active_names.items():
+            self.set_value(
+                self.main_window.config_dict, active_list_key, sorted(names)
+            )
+
+    def save_settings(self):
+        """Save the current settings to the loaded config file."""
+        import toml
+
+        self._sync_settings_inputs_to_config_dict()
+
+        if self.main_window.config_file != "":
+            with open(self.main_window.config_file, "w") as f:
+                toml.dump(self.main_window.config_dict, f)
+            self.main_window.print_log(
+                f"Settings saved to {self.main_window.config_file}"
+            )
+
+            # Refresh the currently-displayed settings panel if one is open, so
+            # newly added registry entries immediately appear in dropdowns etc.
+            if self.main_window._last_settings_view:
+                # Capture the currently active tab index
+                from PySide6.QtWidgets import QTabWidget
+
+                current_tab_index = None
+                if self.main_window.settings_layout.count() > 0:
+                    widget = self.main_window.settings_layout.itemAt(0).widget()
+                    if isinstance(widget, QTabWidget):
+                        current_tab_index = widget.currentIndex()
+
+                # Replay the last view to refresh
+                if self.main_window._last_settings_view[0] == "full":
+                    self.display_full_settings()
+                elif self.main_window._last_settings_view[0] == "action":
+                    _, action, checked_ids = self.main_window._last_settings_view
+                    self.display_settings(action, checked_ids)
+
+                # Restore the tab index if possible
+                if current_tab_index is not None:
+                    widget = self.main_window.settings_layout.itemAt(0).widget()
+                    if isinstance(widget, QTabWidget):
+                        if 0 <= current_tab_index < widget.count():
+                            widget.setCurrentIndex(current_tab_index)
+        else:
+            self.main_window.print_log("Settings not saved, please choose a config file")
+
+    def _render_settings_tabs(self, settings_by_section):
+        """Render settings_by_section dict into tabbed layout.
+
+        Clears settings_layout and builds QTabWidget with one tab per section,
+        populating form inputs via settings_factory.build_tab_form.
+        Shared with both display_settings and display_full_settings.
+        """
+        from PySide6.QtWidgets import QTabWidget, QVBoxLayout, QLabel
+
+        # Flush any pending edits into config_dict before the widgets holding
+        # them are destroyed below, so switching tabs never silently drops
+        # unsaved changes (e.g. newly added registry rows).
+        if self.main_window.settings_inputs:
+            self._sync_settings_inputs_to_config_dict()
+
+        while self.main_window.settings_layout.count():
+            child = self.main_window.settings_layout.takeAt(0)
+            if child.widget():
+                child.widget().deleteLater()
+
+        self.main_window.settings_inputs = {}
+
+        # If no sections, add a message and return
+        if not settings_by_section:
+            self.main_window.settings_layout.addWidget(QLabel("No settings available"))
+            self.main_window.settings_layout.addStretch()
+            return
+
+        # Create a tabbed interface with one tab per section
+        tabs = QTabWidget()
+
+        # Iterate through sections in order (dict preserves insertion order in Python 3.7+)
+        for section, settings_list in settings_by_section.items():
+            # Create a scroll area and container for this section's fields
+            tab_container = QWidget()
+            tab_form = QFormLayout(tab_container)
+            tab_form.setContentsMargins(8, 8, 8, 8)  # Padding on all sides
+
+            # Create form_context for multi_file/path_map dynamic row insertion
+            form_context = {"form": tab_form}
+
+            # Build rows in the form, handling grouping via group_name metadata
+            self.build_tab_form(
+                tab_form, settings_list, form_context=form_context
+            )
+
+            # Add stretch at the end of the section (QFormLayout doesn't auto-stretch rows)
+            tab_form.setRowWrapPolicy(QFormLayout.WrapLongRows)
+            # Push content to top by adding a stretch row at the end
+            tab_form.addItem(QVBoxLayout())
+
+            # Add this section as a tab (capitalize section name for display)
+            tabs.addTab(tab_container, section.capitalize())
+
+        self.main_window.settings_layout.addWidget(tabs)
+        self.main_window.settings_layout.addStretch()
+
+    def display_settings(self, action, checked_ids):
+        """Display the relevant settings based on the action being used."""
+        self.main_window._last_settings_view = ("action", action, checked_ids)
+        relevant_settings = self.get_relevant_settings(action, checked_ids)
+        self._render_settings_tabs(relevant_settings)
+
+    def display_full_settings(self):
+        """Display all fixed-schema sections in the settings panel."""
+        self.main_window._last_settings_view = ("full",)
+        all_settings = self.get_all_settings()
+        self._render_settings_tabs(all_settings)
